Move debug drawing to Video
